01_agent_story_writer/pyproject.toml
[project]
name = "somple-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.8",
    "python-dotenv>=1.1.0",
]
